requirements.txt
boto3==1.4.2
botocore==1.4.82
configparser==3.5.0
docutils==0.12
jmespath==0.9.0
pep8==1.7.0
Pillow==3.4.1
python-dateutil==2.6.0
s3transfer==0.1.9
six==1.10.0
